refactor export dataset to notebook test
